+++ .eslintrc.js
@@ -18,6 +18,20 @@ module.exports = {
 	],
 	rules: {
 		'@typescript-eslint/no-var-requires': 0,
+		'@typescript-eslint/typedef': [
+			"warn",
+			{
+				arrayDestructuring: false,
+				arrowParameter: false,
+				memberVariableDeclaration: true,
+				objectDestructuring: false,
+				parameter: true,
+				propertyDeclaration: true,
+				variableDeclaration: false,
+				variableDeclarationIgnoreFunction: true
+			}
+		],
+		'@typescript-eslint/no-inferrable-types': 0
 	},
 	ignorePatterns: ['node_modules'],
 	overrides: [